backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.31.0
psycopg2-binary
alembic==1.13.3

# Settings
pydantic-settings==2.6.1
pydantic==2.7.4

# OpenAI
openai==1.54.0

# Telegram
python-telegram-bot==21.6

# HTTP Client
httpx==0.27.0

# Scheduler
APScheduler==3.10.4

# Document Processing
pdfplumber==0.11.4
python-docx==1.1.2

# Logging
structlog==24.4.0

# Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
slowapi==0.1.9

# Utilities
python-dateutil==2.9.0
aiofiles==24.1.0

# Web scraping
requests
beautifulsoup4
openpyxl==3.1.5
